+++ packages/typespec-rust/CHANGELOG.md
@@ -22,6 +22,7 @@
 ### Other Changes
 
 * Emit `#[allow(clippy::too_many_arguments)]` on methods that contain seven or more arguments.
+* Emit doc comments for synthesized nullable types.
 
 ## 0.24.1 (2025-10-15)